WIP: split out highlight.c for syntax.c
